lambda/requirements.txt
torch==2.1.0
transformers==4.36.0
python-dotenv==1.0.0
requests==2.31.0
urllib3==2.0.7 